+++ src/Controls/tests/UITests/TestCategory.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Maui.AppiumTests
+{
+	public static class TestCategory
+	{
+		public const string Layout = nameof(Layout);
+	}
+}
